refactor(platform): move system permissions impl
