+++ docs/architecture/platform.md
@@ -20,6 +20,7 @@
 | `detect_system_proxy() -> Option<String>` | `OnceLock` 进程级缓存；优先 env vars（`HTTPS_PROXY` / `HTTP_PROXY` / `ALL_PROXY` 大小写），macOS 读 `scutil --proxy`，Linux / BSD 再试 GNOME `gsettings` 与 KDE `kreadconfig6` / `kreadconfig5` | 读 `HKCU\Software\Microsoft\Windows\CurrentVersion\Internet Settings`，`OnceLock` 进程级缓存避免每次构建 client 都重读注册表；解析 `ProxyEnable` + `ProxyServer`，支持 `http=...;https=...` 协议列表（优先 https） |
 | `current_location() -> Option<(f64, f64)>` | macOS 走 CoreLocation；其他 Unix 返回 `None`，业务层继续 IP 定位降级 | 返回 `None`，业务层继续 IP 定位降级 |
 | `pdfium_library_candidates() -> &'static [&'static str]` | macOS 返回 Homebrew / `/usr/local` dylib 候选；其他 Unix 返回常见 `.so` 候选 | 返回 `pdfium.dll` |
+| `system_permissions_*` | macOS 走 TCC / framework 原生检查与授权触发；非 macOS 返回 unsupported / NotApplicable | 返回 unsupported / NotApplicable |
 | `default_shell_command(cmdline) -> std::process::Command` | `Command::new("sh").arg("-c").arg(cmdline)` | `Command::new("cmd").raw_arg("/C").raw_arg(cmdline)` —— `raw_arg` 跳过 std 自动加引号，保留 `/C` 后续整段命令的原始语义 |
 | `default_shell_command_tokio(cmdline)` | 同上 std 版，返回 `tokio::process::Command` | 同上 std 版，返回 `tokio::process::Command` |
 | `os_version_string() -> String` | macOS 优先 `sw_vers -productVersion` → `"macOS 14.2.1"` 形态；其他 Unix 走 `sysinfo::System::long_os_version()` 兜底；都失败时 `"unknown"` | `sysinfo::long_os_version()` + `kernel_version()` 拼成 `"Windows 11 (26100)"` 形态；都缺失时 `"Windows (unknown build)"` |
@@ -68,6 +69,7 @@
 | `detect_system_proxy` | [`provider/proxy.rs`](../../crates/ha-core/src/provider/proxy.rs) LLM 出站代理；[`docker/proxy.rs`](../../crates/ha-core/src/docker/proxy.rs) Docker 容器代理注入 |
 | `current_location` | [`weather.rs`](../../crates/ha-core/src/weather.rs) 天气自动定位：系统精确定位失败后降级 IP 定位 |
 | `pdfium_library_candidates` | [`file_extract.rs`](../../crates/ha-core/src/file_extract.rs) PDF 渲染 fallback 动态库查找 |
+| `system_permissions_*` | [`permissions.rs`](../../crates/ha-core/src/permissions.rs) v2 系统权限目录的 OS 原生检查 / 请求入口 |
 | `default_shell_command_tokio` | [`tools/exec.rs`](../../crates/ha-core/src/tools/exec.rs) 工具 shell 命令执行 |
 | `os_version_string` | [`agent/errors.rs`](../../crates/ha-core/src/agent/errors.rs) 错误报告 / 诊断；`self_diagnosis` 日志 |
 | `write_secure_file` | [`mcp/credentials.rs`](../../crates/ha-core/src/mcp/credentials.rs) MCP OAuth token 凭据 0600 原子落盘（**当前唯一调用方**）。注意：主 LLM OAuth `oauth.rs::save_token()` 当前直接用 `std::fs::write` 写 `~/.hope-agent/credentials/auth.json`，**未走** `write_secure_file`——见下文「已知缺口」 |
@@ -85,6 +87,7 @@
 
 | 文件 | 职责 |
 |---|---|
-| [`crates/ha-core/src/platform/mod.rs`](../../crates/ha-core/src/platform/mod.rs) | 门面：12 个 `pub fn` 入口 + 跨平台 doc 注释，编译期按 `#[cfg(unix)]` / `#[cfg(windows)]` route 到对应 impl |
+| [`crates/ha-core/src/platform/mod.rs`](../../crates/ha-core/src/platform/mod.rs) | 门面：系统级 `pub fn` 入口 + 跨平台 doc 注释，编译期按 `#[cfg(unix)]` / `#[cfg(windows)]` route 到对应 impl |
 | [`crates/ha-core/src/platform/unix.rs`](../../crates/ha-core/src/platform/unix.rs) | Unix 实现：`libc::kill` / `sh -c` / `OpenOptions::mode(0o600)` / `sw_vers` 兜底 / `chromiumoxide` 走自己的 which |
 | [`crates/ha-core/src/platform/windows.rs`](../../crates/ha-core/src/platform/windows.rs) | Windows 实现：`taskkill /F /T` / `cmd /C raw_arg` / NTFS DACL 继承 / winreg 读 Internet Settings + `OnceLock` 缓存 / `%ProgramFiles%` 三路扫 Chrome |
+| [`crates/ha-core/src/platform/system_permissions.rs`](../../crates/ha-core/src/platform/system_permissions.rs) | 系统权限 OS 实现：macOS TCC / framework 原生权限检查与 prompt；非 macOS 明确 unsupported |
