+++ docs/plans/review-followups.md
@@ -54,17 +54,16 @@
 - **来源**：2026-05-01 跨平台兼容性修复 PR（`claude/cross-platform-compatibility-check-qBENn`）
 - **现象**：本期 PR 只修了"主路径在非 macOS / 非 Tauri 直接走不通"的两个硬伤（Skills 目录选择 + Ollama 失败 UX）。仓库里仍有大量 `#[cfg(target_os = "linux")]` / `#[cfg(target_os = "macos")]` 散落在业务代码而非 `crates/ha-core/src/platform/` 门面下，违反 AGENTS.md「优先用 `#[cfg(unix)]` / `#[cfg(windows)]`，少写 `target_os = "linux"`；新增跨平台原语统一放 `crates/ha-core/src/platform/`」规则。重灾区：
   - [`crates/ha-core/src/service_install.rs`](../../crates/ha-core/src/service_install.rs)：~30 处 `target_os = "macos"` / `"linux"` 分支硬编码（launchd plist / systemd unit 业务逻辑应进 `platform::service` 子模块）
-  - [`crates/ha-core/src/permissions.rs:56,318`](../../crates/ha-core/src/permissions.rs)：macOS-only TCC 权限申请，非 macOS 全 stub——能走通但应迁到 `platform::permissions` 门面
 - **2026-05-20 已处理**：系统代理探测已收敛到 [`platform::detect_system_proxy`](../../crates/ha-core/src/platform/mod.rs)：macOS 走 `scutil --proxy`，Linux / BSD 走 env → GNOME `gsettings` → KDE `kreadconfig6/5`，provider / Docker 不再各自维护 `scutil` 分支
 - **2026-05-20 已处理**：天气自动定位已改走 [`platform::current_location`](../../crates/ha-core/src/platform/mod.rs)，[`weather.rs`](../../crates/ha-core/src/weather.rs) 不再直接带 macOS CoreLocation cfg；非 macOS 仍按原行为降级 IP 定位
 - **2026-05-20 已处理**：PDFium fallback 动态库候选路径已改走 [`platform::pdfium_library_candidates`](../../crates/ha-core/src/platform/mod.rs)，[`file_extract.rs`](../../crates/ha-core/src/file_extract.rs) 不再直接带 OS cfg；原条目里提到的 Office `textutil` / `libreoffice` 分支已不存在，当前 Office 提取是纯 Rust 路径
-- **为什么留**：本期 PR 范围聚焦"完全走不通"的两个硬伤（用户已经在反馈），上面这些都不是 blocker：permissions 在非 macOS 已有 NotApplicable 降级。逐个迁到 `platform/` 是大块结构性重构，不能跟修主路径混在一个 PR
+- **2026-05-20 已处理**：系统权限原生实现已迁到 [`platform/system_permissions.rs`](../../crates/ha-core/src/platform/system_permissions.rs)，[`permissions.rs`](../../crates/ha-core/src/permissions.rs) 只保留权限目录、响应 shaping 与 legacy 映射；非 macOS 仍按原行为返回 unsupported / NotApplicable
+- **为什么留**：`service_install.rs` 是最大块，launchd / systemd 文本、安装/卸载/状态查询耦合在一个业务文件里，需要单独拆 `platform::service`，不适合和其它小门面迁移混在一起
 - **改的话要做什么**：
-  1. 先做 audit：grep 全 `target_os =` 出现位置，按"业务文件 vs platform 门面"分两堆
-  2. 对每个业务文件里的分支，判断是该 (a) 整个函数迁到 `platform/{macos,linux,windows}.rs` 然后 `crate::platform::xxx()` 调用，还是 (b) 改成 `cfg(unix)` 让 macOS+Linux+BSD 共享路径
-  3. `service_install.rs` 拆成 `platform::service::{install,uninstall,status}` + 各 OS 实现文件——是最大的一块
+  1. `service_install.rs` 拆成 `platform::service::{install,uninstall,status}` + 各 OS 实现文件
+  2. 保留 public API 不变，让 CLI / updater service control 继续调同一入口
 - **影响面**：全是"已经能跑但不够 OS-native"——非 macOS 用户拿到的是降级体验或不支持提示。无安全 / 数据正确性问题，但跨平台口碑会被这些细节拖累
-- **触发时机建议**：可以拆成 2 个独立小 PR 渐进推进（`service_install` / `permissions` 各一个），每个独立可 review；或者下次有 Windows / Linux 用户报某个具体子系统不能用时，趁势把对应那块迁到 platform 门面
+- **触发时机建议**：最后拆 `service_install`；或者下次有 Windows / Linux 用户报服务安装/状态查询不能用时，趁势把对应那块迁到 platform 门面
 
 ### F-027 9 个语言 `settings.approvalPanel` block 是英文 verbatim fallback
